+++ pulsar/apps/wsgi/server.py
@@ -290,8 +290,8 @@ class HttpServerResponse(ProtocolConsumer):
     SERVER_SOFTWARE = pulsar.SERVER_SOFTWARE
     ONE_TIME_EVENTS = ProtocolConsumer.ONE_TIME_EVENTS + ('on_headers',)
 
-    def __init__(self, wsgi_callable, cfg, server_software=None):
-        super(HttpServerResponse, self).__init__()
+    def __init__(self, wsgi_callable, cfg, server_software=None, loop=None):
+        super(HttpServerResponse, self).__init__(loop=loop)
         self.wsgi_callable = wsgi_callable
         self.cfg = cfg
         self.parser = http_parser(kind=0)